Create Event model wrapper; add some policies and actions; reorganize
