+++ authorization/attributes/actions/__init__.py
@@ -1,1 +1,3 @@
 from .user import *
+from .group import *
+from .event import *
