+++ authorization/attributes/actions/user/calendar.py
@@ -1,19 +1,19 @@
-from .binary_user_action import BinaryUserAction
+from ..bounded_actions import BinaryUserAction
 from authorization import policies
 
 
 class ViewUserCalendar(BinaryUserAction):
     policies = [
-        # A user can view his or her own profile.
+        # A user can view his or her own calendar.
         policies.miscellaneous.SubjectIsResource,
         
-        # Users who are contacts can view each other's profile.
+        # Users who are contacts can view each other's calendar.
         policies.user.relationship.UsersAreContacts
     ]
 
 
 class EditUserCalendar(BinaryUserAction):
     policies = [
-        # A user can edit his or her own profile.
+        # A user can edit his or her own calendar.
         policies.miscellaneous.SubjectIsResource
     ]
