+++ authorization/attributes/resources/__init__.py
@@ -1,1 +1,2 @@
 from .user import User
+from .group import Group
