+++ authorization/information/__init__.py
@@ -1,2 +1,3 @@
 from .user import User
 from .group import Group
+from .event import Event
