+++ authorization/policies/__init__.py
@@ -1,2 +1,3 @@
-from .miscellaneous import *
 from .user import *
+from .group import *
+from .miscellaneous import SubjectIsResource
