--- authorization/policies/miscellaneous/subject_is_resource.py
@@ -1,7 +1,0 @@
-from authorization.decision import Policy
-
-
-class SubjectIsResource(Policy):
-    @classmethod
-    def evaluate(cls, request):
-        return request.subject == request.resource
