+++ authorization/policies/user/__init__.py
@@ -1,1 +1,1 @@
-from .relationship import *
+from .relationship import UsersAreContacts, UsersHaveACommonGroup
